backend_dl/requirements.txt
# backend_CNN/requirements.txt

fastapi
numpy
opencv-python
pillow
python-multipart
uvicorn
pandas
request